+++ machine_learning_library/classification.h
@@ -9,7 +9,7 @@ typedef struct {
     network* net;
 } ClassificationNetwork;
 
-ClassificationNetwork* ClassificationNetwork_create(int layerAmount, int* layersSize, int* input_shape, int input_dim, ActivationType* activations, double learnningRate, LossType lossFunction, char** class_names, int num_classes, Tensor* classes, LayerType type);
+ClassificationNetwork* ClassificationNetwork_create(int layerAmount, int* layersSize, int* input_shape, int input_dim, ActivationType* activations, float  learnningRate, LossType lossFunction, char** class_names, int num_classes, Tensor* classes, LayerType type);
 ClassificationNetwork* ClassificationNetwork_create_net(network* net, char** class_names, int num_classes, Tensor* classes);
 Tensor* one_hot_encode(int num_classes);
 void classification_info_free(ClassificationNetwork* info);
